feat(tests): swap to Verify, for better management of snapshots
