+++ Brotli.NET/.gitattributes
@@ -56,3 +56,7 @@
 *.rtf  diff=astextplain
 *.RTF  diff=astextplain
 
+*.verified.txt text eol=lf working-tree-encoding=UTF-8
+*.verified.xml text eol=lf working-tree-encoding=UTF-8
+*.verified.json text eol=lf working-tree-encoding=UTF-8
+
