+++ Brotli.NET/Brotli.NET.Tests/Brotli.NET.Tests.csproj
@@ -4,6 +4,7 @@
 	<TargetFramework>net7.0</TargetFramework>
 
     <IsPackable>false</IsPackable>
+	<ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
@@ -14,6 +15,8 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
     <PackageReference Include="Shouldly" Version="4.3.0" />
+
+	<PackageReference Include="Verify.Xunit" Version="29.1.0" />
     <PackageReference Include="xunit" Version="2.9.3" />
     <PackageReference Include="xunit.runner.visualstudio" Version="3.0.2">
       <PrivateAssets>all</PrivateAssets>
